Adapted to psoc6-evaluationkit-062S2

+++ rt-thread/components/libc/posix/io/poll/select.c
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2006-2022, RT-Thread Development Team
+ * Copyright (c) 2006-2021, RT-Thread Development Team
  *
  * SPDX-License-Identifier: Apache-2.0
  *
@@ -100,7 +100,7 @@ int select(int nfds, fd_set *readfds, fd_set *writefds, fd_set *exceptfds, struc
     /* Convert the timeout to milliseconds */
     if (timeout)
     {
-        msec = timeout->tv_sec * 1000 + timeout->tv_usec / 1000;
+        msec = (int)timeout->tv_sec * 1000 + (int)timeout->tv_usec / 1000;
     }
     else
     {
